Moved all recommended/ apis to deprecated controller, these haven't been used in 2+ years.
